Game v1.6 (animation, character color, lighten - darken, order of characters, instructions and parameters to process from a file)
The character performs a pre-made animation (e.g. shiver). When he doesn't speak, he's in the shadows, the one who speaks will be in the spotlight. The character can move his face sideways. The color of the entire image can be changed. If there are several characters next to each other so that they cover each other, the order can be changed. The instructions are processed from a file, they can be interpreted with the help of parameters.